rename wait_for_read to wait_until_ready
