+++ twitchchat/src/runner/mod.rs
@@ -6,7 +6,7 @@ mod error;
 mod identity;
 mod timeout;
 mod wait_for;
-pub(crate) mod wait_for_ready;
+pub(crate) mod wait_until_ready;
 
 pub use capabilities::Capabilities;
 pub use error::Error;
@@ -16,11 +16,11 @@ pub use timeout::{idle_detection_loop_sync, TIMEOUT, WINDOW};
 pub use timeout::{Activity, ActivityReceiver, ActivitySender};
 
 pub use wait_for::{wait_for_sync, Event};
-pub use wait_for_ready::wait_until_ready_sync;
+pub use wait_until_ready::wait_until_ready_sync;
 
 cfg_async! {
     pub use timeout::idle_detection_loop;
-    pub use wait_for_ready::wait_until_ready;
+    pub use wait_until_ready::wait_until_ready;
 
     cfg_writer! {
         pub use timeout::respond_to_idle_events;
